backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
PyYAML
LogicBank
pytest
httpx
git+https://github.com/thomaxxl/safrs.git
